# internal/management/openapi.yaml
openapi: 3.0.3
info:
  title: Rate Limit Management API
  version: 1.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: |
    Control interface of the rate limit service: read what is enforced,
    inspect and reset live counters, simulate a request, and tap the decision
    audit stream.

    Principles:
    * The API mutates runtime state only (counters, audit taps). Configuration
      (limits, rules, claim mappings, groups) lives in the RateLimitPolicy of
      the domain, the one custom resource - the GitOps path. There are no
      configuration write endpoints.
    * Reads never charge counters: every read goes through the store's Peek,
      evaluated at cost=1. Judging arbitrary costs is what the simulation
      endpoint is for.
    * Every mutation is audited - a structured audit log line with the
      request id and subject is the durable journal; a Kubernetes Event is
      emitted as an operational signal on top, never as the record.
    * Destructive radius is bounded by construction: DELETE only addresses
      one rule; anything wider goes through the counter-resets action, whose execution step only exists with the single-use confirmation token minted by a mandatory preview - look before you delete is enforced by construction, and the sweep's work is bounded server-side: the walk has a deadline, and a domain runs at most one sweep at a time.
...

...
    the wrong place for a thousand client ids.

...
    Catalog:
...

...
servers:
...
security:
- bearerAuth: []
...
paths:
  /domains:
    get:
      tags:
      - domains
      operationId: listDomains
      summary: List the domains under management
      responses:
        '200':
          description: Every domain the current rule set carries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainList'
              example:
                items:
                - domain: gateway.public
                  ruleSetVersion: 7c31a9f4e0d2
                  blocks: 8
                  rules: 15
                  effectiveKeys:
                  - client
                  - plan
                  - roles
                  - tenant
                  listValuedKeys:
                  - roles
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
      description: 'The entry point of an investigation: which domains the enforced
        rule set carries and how big each one is. Domains are few by design, so
        the listing is not paginated.'
      parameters:
      - $ref: '#/components/parameters/xRequestId'
...
  /simulations:
    post:
      tags:
      - simulations
      operationId: simulateCheck
      summary: Ask what would happen to a request, charging nothing
      description: |
        Runs the full decision pipeline - matching, identity extraction,
        per-rule verdicts - through the store's non-charging read, at the
        cost the body names. The answer is the verdict as of evaluatedAt - nothing is reserved, and parallel traffic can change the real outcome - plus the per-rule breakdown the gateway never sees.

        Identity comes from `identitySource`; the three request forms are
        structural (see SimulationRequest):
        * `merge` (default) - extract from the token, then explicit keys
          override per key. This is the gateway's own semantics; with
          neither present the simulation is anonymous.
        * `token` - extraction only; the form admits no keys field.
        * `keys` - explicit keys only; the form admits no token field.

        The token is write-only input: it is never echoed in the response,
        never written to access or audit logs, never quoted in errors, and
        never reaches the audit stream.

        The domain travels in the body rather than the path: one mutation-free POST path keeps the role model trivial - viewer holds post on exactly this path and never on the mutations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
            examples:
              merge:
                summary: 'Gateway semantics: token extraction with keys overlaid'
                value:
                  domain: gateway.public
                  path: /api/quotes/42
                  method: GET
                  keys:
                    client:
                    - alice
                    tenant:
                    - acme
                    plan:
                    - gold
              tokenOnly:
                summary: Pure token - what extraction alone would judge
                value:
                  domain: gateway.public
                  path: /api/quotes/42
                  method: GET
                  identitySource: token
                  token: eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiJhbGljZSJ9.sig
              anonymous:
                summary: No identity at all - what the unconditional rules say
                value:
                  domain: gateway.public
                  path: /api/login
                  method: POST
      responses:
        '200':
          description: The decision the gateway would have received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
              example:
                allowed: false
                evaluatedAt: '2026-08-24T14:02:09Z'
                refusalReason: rate_limited
                headers:
                  limit: 100
                  remaining: 0
                  retryAfterSeconds: 23.4
                  resetAfterSeconds: 60
                  algorithm: gcra
                  periodSeconds: 60
                rules:
                - id: cascade/everyone
                  mode: enforce
                  allowed: false
                  refusalReason: rate_limited
                  limit: 100
                  remaining: 0
                  retryAfterSeconds: 23.4
                  algorithm: gcra
                  periodSeconds: 60
                - id: quota/daily
                  mode: enforce
                  allowed: true
                  limit: 10000
                  remaining: 8734
                  algorithm: fixedwindow
                  periodSeconds: 86400
                - id: tenants/per-tenant
                  mode: enforce
                  allowed: true
                  limit: 1000
                  remaining: 902
                  algorithm: gcra
                  periodSeconds: 60
                - id: total/all
                  mode: enforce
                  allowed: true
                  limit: 5000
                  remaining: 4310
                  algorithm: gcra
                  periodSeconds: 60
                skips: []
                extractedKeys:
                - client
                - plan
                - tenant
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
        '404':
          description: Unknown domain (RLS-0404).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmfErrorResponse'
              example:
                id: c5d17a94-2e63-4f80-8b19-0a7c4e9f3d26
                code: RLS-0404
                reason: Unknown resource
                message: domain "gateway.typo" is not in the enforced rule set
                status: '404'
                '@type': NC.TMFErrorResponse.v1.0
                meta:
                  requestId: c81d2f96
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: 'Invalid request: an identity form violated (token form without
            a token, keys form with an empty map) or malformed body (RLS-0400).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmfErrorResponse'
              example:
                id: 8c53a0e7-1d49-4b26-9f38-6e2a7d0c4b91
                code: RLS-0400
                reason: Invalid request
                message: identitySource token requires the token field
                status: '400'
                '@type': NC.TMFErrorResponse.v1.0
                meta:
                  requestId: 6b21f4d8
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - $ref: '#/components/parameters/xRequestId'
  /status:
    get:
      tags:
      - meta
      operationId: getStatus
      summary: What this replica is serving right now
      description: |
        Per-replica view: the custom resource status shows the consensus,
        while replicas can diverge - briefly during a rollout, or durably when their last-good histories differ - and this is the endpoint that makes the divergence visible.
      responses:
        '200':
          description: The replica's serving state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusView'
              example:
                replica: ratelimit-6c9d-x2v
                snapshotSwappedAt: '2026-08-24T13:58:41Z'
                ruleSetVersions:
                  gateway.public: 7c31a9f4e0d2
                counterStore:
                  backend: redis at redis:6379
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - $ref: '#/components/parameters/xRequestId'
  /openapi.yaml:
    get:
      tags:
      - meta
      operationId: getOpenAPI
      summary: This specification
      responses:
        '200':
          description: The OpenAPI document.
          content:
            application/yaml:
              schema:
                type: string
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
      description: 'The specification, embedded at build time: what you read here
        is what this binary was built against, not a copy that can drift.'
      parameters:
      - $ref: '#/components/parameters/xRequestId'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A bearer JWT issued by the platform identity provider. Authentication is the perimeter's job: the gateway's auth extension validates the token (signature, expiry, issuer) before the request reaches the service, and the mesh restricts the service's ingress to that gateway. Authorization is the service's own, and its trust boundary is explicit: identity is read from EXACTLY ONE source - the bearer token in the Authorization header, whose signature the gateway has already validated; no auxiliary identity header (X-Forwarded-User and kin) is ever consulted, so none can be spoofed. What makes the unvalidated read safe is a DEPLOYMENT REQUIREMENT, not an assumption: the mesh must restrict the service's ingress to the gateway alone; a deployment that cannot enforce that must validate the token signature in-service instead. The service enforces the role model per path and verb - viewer holds every GET and POST /simulations; operator holds the mutations (DELETE on counters, POST on counter-resets, PUT on the audit selection) plus everything viewer holds. Role names and claim locations are deployment configuration; viewer and operator are the canonical defaults. The subject also feeds the audit journal, the Idempotency-Key scope, and the confirmation-token binding. A request without a bearer token is refused outright (401) - but that is hygiene, not the defense: the service does not validate signatures, so with the ingress requirement broken a forged token would pass. The model's security rests on that requirement, or on in-service validation where the mesh cannot provide it. An invalid token is normally rejected at the gateway, in the perimeter's own error format. Roles cannot narrow to particular axis values: whoever holds viewer sees axis values (client ids), which is a documented property of the access model, not an accident.
  parameters:
...
    xRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 'Caller-supplied request id; it round-trips into the X-Request-Id
        response header and ties the call to the operator log, the audit journal,
        and meta.requestId of error bodies. Absent, the server generates one. Values
        outside the pattern are refused (RLS-0400), never sanitized: the id lands
        verbatim in logs and the audit journal, so only log-safe characters are
        admitted.'
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
  responses:
    Forbidden:
      description: Access denied (RLS-0403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TmfErrorResponse'
          example:
            id: e4b19c6d-2a53-4f70-8d14-6b9e0c3a7f28
            code: RLS-0403
            reason: Access denied
            message: subject support-viewer lacks the operator role for DELETE /ratelimit/v1/domains/gateway.public/counters
            status: '403'
            '@type': NC.TMFErrorResponse.v1.0
            meta:
              requestId: 9d1f4c2e
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
    Error:
      description: |
        Any error. The body is NC.TMFErrorResponse.v1.0; branch on `code`
        (see the catalog in the API description), read the HTTP status from
        `status`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TmfErrorResponse'
          example:
            id: 3f9a7b2c-8e15-4d60-a2c9-1b4f6d8e0a37
            code: RLS-0404
            reason: Unknown resource
            message: domain "gateway.typo" is not in the enforced rule set
            status: '404'
            '@type': NC.TMFErrorResponse.v1.0
            meta:
              requestId: b21c9f30
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
    Unauthorized:
      description: 'Authentication required (RLS-0401): the request arrived without
        a bearer token. An invalid or expired one is normally rejected at the gateway,
        in the perimeter''s own format.'
      headers:
        WWW-Authenticate:
          description: Always Bearer - the API expects the platform IdP bearer token.
          required: true
          schema:
            type: string
            example: Bearer
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TmfErrorResponse'
          example:
            id: d82f5c30-6a19-4e47-b0d8-3c7f1e9a5b62
            code: RLS-0401
            reason: Authentication required
            message: no bearer token on the request; call through the platform gateway
            status: '401'
            '@type': NC.TMFErrorResponse.v1.0
            meta:
              requestId: 5c19e8f4
  schemas:
    TmfErrorResponse:
      type: object
      description: |
        NC.TMFErrorResponse.v1.0, produced with
        qubership-core-lib-go-error-handling. `id` identifies the error
        instance (quote it when asking why a call failed), `code` is the
        stable contract to branch on, `reason` is the code's title, `message`
        carries the context detail.
      required:
      - id
      - code
      - reason
      - message
      - '@type'
      - status
      - meta
      properties:
        id:
          type: string
          format: uuid
          description: Unique id of this error instance - what a person quotes when
            asking why a call failed.
        code:
          type: string
          pattern: ^([A-Z]+-)+\d{4}$
          example: RLS-0404
          description: The stable machine contract to branch on; see the catalog
            in the API description.
        reason:
          type: string
          example: Unknown resource
          description: The error code's title - constant per code, never per instance.
        message:
          type: string
          example: domain "gateway.typo" is not in the enforced rule set
          description: Human-readable detail of this instance; free to change, never
            a contract.
        referenceError:
          type: string
          description: URL of documentation for the error, when one exists.
        status:
          type: string
          example: '404'
          description: The HTTP status code, carried as a string per the TMF envelope;
            always present, so a client reading the body alone knows the outcome
            class.
        source:
          description: The request element the error originates from, when one is
            identifiable.
        meta:
          type: object
          additionalProperties: true
          description: Carries requestId; for validation errors, fields; on every
            409 the service forms, conflictType; for a bulk reset that failed after
            acceptance, partialReset.
          required:
          - requestId
          properties:
            requestId:
              type: string
              minLength: 1
              maxLength: 128
              description: The request id of the failing call - the X-Request-Id
                value; quote it together with id when asking what happened.
              pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
            fields:
              type: array
              items:
                type: string
                description: One request field at fault.
              description: Request fields the validation error is about.
            partialReset:
              $ref: '#/components/schemas/PartialReset'
            conflictType:
              type: string
              enum:
              - command_mismatch
              - stale_confirmation
              - stale_rule_set
              - sweep_in_flight
              - stale_if_match
              description: 'Which conflict this 409 is, because each recovers differently:
                command_mismatch - new Idempotency-Key or the original body; stale_confirmation
                - new preview; stale_rule_set - re-read the enforced set and repeat;
                sweep_in_flight - retry after Retry-After; stale_if_match - re-read
                the ETag, merge, retry. Present on every 409 the service forms.'
        errors:
          type: array
          items:
            type: object
            description: One nested error of the aggregate.
          description: Nested component errors of an aggregate failure.
        '@type':
          type: string
          enum:
          - NC.TMFErrorResponse.v1.0
          description: The TMF envelope type marker.
        '@schemaLocation':
          type: string
          description: URL of the schema describing this error structure.
    DomainList:
      type: object
      description: Domains are few by design; the listing is not paginated.
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DomainSummary'
          description: Every domain of the enforced rule set.
...
    SimulationRequest:
      description: |
        Three structural forms, discriminated by identitySource. merge (the
        default when the field is absent) mirrors the gateway: extraction
        from the token with explicit keys overlaid per key, and with
        neither present the simulation is anonymous - a legitimate question
        (what do the unconditional rules say). The pure forms admit exactly
        their own source, so what was judged is never ambiguous. The
        runtime enforces no additional structural constraints.
      oneOf:
      - $ref: '#/components/schemas/SimulationMerge'
      - $ref: '#/components/schemas/SimulationToken'
      - $ref: '#/components/schemas/SimulationKeys'
...
    SimulationResponse:
      description: The decision the gateway would have received, plus the per-rule
        breakdown the gateway never sees.
      oneOf:
      - $ref: '#/components/schemas/SimulationAdmitted'
      - $ref: '#/components/schemas/SimulationRateLimited'
      - $ref: '#/components/schemas/SimulationCapacityExceeded'
...
    StatusView:
      type: object
      required:
      - replica
      - counterStore
      properties:
        replica:
          type: string
          description: This pod's name.
        snapshotSwappedAt:
          type: string
          format: date-time
          description: When the serving rule set last changed on this replica.
        ruleSetVersions:
          type: object
          additionalProperties:
            type: string
            description: One domain's rule-set version.
          description: Per-domain rule-set versions of the serving snapshot - what
            GET /rules reports for each domain; compare across replicas to see rollout
            skew. Absent until the first snapshot lands.
        counterStore:
          type: object
          required:
          - backend
          properties:
            backend:
              type: string
              example: redis at redis:6379
              description: Human-readable backend description, as the startup log
                prints it.
          description: Where the counters live.
      description: One replica's serving state - the per-replica view the custom
        resource status cannot give.
...
  headers:
    XRequestId:
      description: 'The request id of this call: a caller-supplied X-Request-Id
        round-tripped, or a generated one when the caller sent none. The same id
        runs through the operator log, the audit journal, and meta.requestId of
        every error body.'
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$